requirements.txt
speedtest-cli
pandas
matplotlib
numpy